Refactor babylon and jscodeshift (reuse more code)
