+++ registry/tools/babylon/v6.js
@@ -1,12 +1,9 @@
-import React from 'react';
 import base from './base';
-import defaultParserInterface from '../../utils/defaultESTreeParserInterface';
 import pkg from 'babylon/package.json';
 import {parse} from 'babylon';
-import SettingsRenderer from '../../utils/SettingsRenderer';
 
 const ID = 'babylon';
-export const defaultOptions = {
+const defaultOptions = {
   sourceType: 'module',
   allowImportExportEverywhere: false,
   allowReturnOutsideFunction: false,
@@ -26,7 +23,7 @@ export const defaultOptions = {
   ],
 };
 
-export const parserSettingsConfiguration = {
+const settingsConfiguration = {
   fields: [
     ['sourceType', ['module', 'script']],
     'allowReturnOutsideFunction',
@@ -45,45 +42,17 @@ export const parserSettingsConfiguration = {
 };
 
 export default {
-  ...defaultParserInterface,
   ...base,
 
   id: ID,
   displayName: ID,
   version: pkg.version,
   homepage: pkg.homepage,
-  locationProps: new Set(['loc', 'start', 'end']),
+  defaultOptions,
+  settingsConfiguration,
 
   load() {
     return Promise.resolve({parse});
   },
 
-  parse(babylon, code, options) {
-    return babylon.parse(code, {...defaultOptions, ...options});
-  },
-
-  getNodeName(node) {
-    switch (typeof node.type) {
-      case 'string':
-        return node.type;
-      case 'object':
-        return `Token (${node.type.label})`;
-    }
-  },
-
-  nodeToRange(node) {
-    if (typeof node.start !== 'undefined') {
-      return [node.start, node.end];
-    }
-  },
-
-  renderSettings(parserSettings, onChange) {
-    return (
-      <SettingsRenderer
-        settingsConfiguration={parserSettingsConfiguration}
-        parserSettings={{...defaultOptions, ...parserSettings}}
-        onChange={onChange}
-      />
-    );
-  },
 };
